refactor: move styles into own directory
